completed step 7 ... Nothing happened with CommentRepository for some reason think i might try to encourage myself to look into it ;)

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -9,6 +9,7 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         private readonly IPostRepository _postRepository;
         private readonly IBlogRepository _blogRepository;
         private readonly IUserService _userService;
+        private string currentUserId;
 
         public PostService(IPostRepository postRepository, IBlogRepository blogRepository, IUserService userService)
         {
@@ -23,7 +24,18 @@ namespace CS321_W5D2_BlogAPI.Core.Services
             //     Use the _userService to get the current users id.
             //     You may have to retrieve the blog in order to check user id
             // TODO: assign the current date to DatePublished
-            return _postRepository.Add(newPost);
+            var currentUserId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(newPost.BlogId);
+            if(currentUserId == blog.UserId)
+            {
+                newPost.DatePublished = DateTime.Now; // or DateTime.Today
+                return _postRepository.Add(newPost);
+            }
+            else
+            {
+                throw new ApplicationException("Your not allowed in this blog"); 
+                // Exception is the very top ... ApplicationException is a more centered exception                      
+            }
         }
 
         public Post Get(int id)
@@ -45,13 +57,34 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         {
             var post = this.Get(id);
             // TODO: prevent user from deleting from a blog that isn't theirs
-            _postRepository.Remove(id);
+            var currentId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(post.BlogId);
+            if (currentId == blog.UserId)
+            {
+                _postRepository.Remove(id);
+            }
+            else
+            {
+                throw new Exception("You dont have permission!")
+            }
+
         }
 
         public Post Update(Post updatedPost)
         {
             // TODO: prevent user from updating a blog that isn't theirs
-            return _postRepository.Update(updatedPost);
+            var currentId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(updatedPost.BlogId);
+            if(currentUserId == blog.UserId)
+            {
+                return _postRepository.Update(updatedPost)   
+            }
+            else
+            {
+                throw new Exception("Nothingness");
+            }
+
+
         }
 
     }
